Preparing unstable channel

+++ examples/M5Stack-SD-Menu/SD-Apps/gen-menu.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 
-arduino --preserve-temp-files --verbose-build --verify $PWD/examples/$EXAMPLE/$EXAMPLE.ino &>/dev/null
-find /tmp -name \*.partitions.bin -exec rm {} \; #
-find /tmp -name \*.bin -exec mv {} $M5_SD_BUILD_DIR/TobozoLauncher.bin \; #
-cp $M5_SD_BUILD_DIR/TobozoLauncher.bin $M5_SD_BUILD_DIR/menu.bin
+export inofile=$SDAPP_FOLDER/../$EXAMPLE.ino
+export outfile=$SDAPP_FOLDER/../downloader.h
 
-export outfile=$PWD/examples/$EXAMPLE/downloader.h
+[ -f $inofile ] && echo "Compiling $inofile"
+
+[ -f $inofile ] && arduino --preserve-temp-files --verbose-build --verify $PWD/examples/$EXAMPLE/$EXAMPLE.ino &>/dev/null
+[ -f $inofile ] && find /tmp -name \*.partitions.bin -exec rm {} \; #
+[ -f $inofile ] && find /tmp -name \*.bin -exec mv {} $M5_SD_BUILD_DIR/TobozoLauncher.bin \; #
+[ -f $inofile ] && cp $M5_SD_BUILD_DIR/TobozoLauncher.bin $M5_SD_BUILD_DIR/menu.bin
+
+[ -f $inofile ] && echo "Compiling $outfile"
+
 [ -f $outfile ] && sed -i -e 's/"\/sd-updater";/"\/sd-updater\/unstable"/g' $outfile
 [ -f $outfile ] && arduino --preserve-temp-files --verbose-build --verify $PWD/examples/$EXAMPLE/$EXAMPLE.ino &>/dev/null
 [ -f $outfile ] && find /tmp -name \*.partitions.bin -exec rm {} \; #
